Add offline handling
