+++ src/components/index.js
@@ -24,6 +24,7 @@ export * from './RadioListView'
 export * from './ReferencePreview'
 export * from './Routes'
 export * from './Search'
+export * from './SearchOfflinePanel'
 export * from './Settings'
 export * from './Softkey'
 export * from './Table'
