+++ style/feedback.less
@@ -73,6 +73,21 @@
 			}
 		}
 	}
+
+	.search-offline-panel {
+		width: 100vw;
+		height: calc( 100vh - @softkeysHeight - 41px );
+		background-color: #fff;
+		display: flex;
+
+		.search-offline-content {
+			margin: auto;
+
+			.message {
+				font-size: 17px;
+			}
+		}
+	}
 }
 
 .confirmation-popup {
